Update script-base.js
Fix of issue #255

- Respect 'includeREquireJS'-Option in yo-rc.json
- Remove wrong and not longer necessary implementation of 'checkIfUsingRequireJS'

Reason:

Existing check did not respect configuration option. It checks if a file at a hardcoded subpath exists or not. This might deliver a result which looks correct as long as the user did not change the default app-path from 'app' to e.g. 'myapp' - but in fact it's not a requirejs detection and it did not respect the 'includeRequireJS'-Option.